refactor find_cloned_card to separate class
